Adds multiple authors condition
