+++ src/components/publishing/related_articles/related_articles_canvas.tsx
@@ -24,7 +24,7 @@ const RelatedArticlesCanvas: React.SFC<RelatedArticlesCanvasProps> = props => {
   } else {
     return (
       <RelatedArticlesContainer>
-        <Title>Further Reading <VerticalSpan>{vertical.name}</VerticalSpan></Title>
+        <Title>Further Reading in <VerticalSpan>{vertical.name}</VerticalSpan></Title>
         <ArticlesWrapper>
           {_.map(articles, (article, i) => {
             return <RelatedArticleFigure article={article} key={`related-article-figure-${i}`} />
@@ -49,19 +49,7 @@ const Title = styled.div`
   margin-bottom: 30px;
 `
 const VerticalSpan = styled.span`
-  margin-left: 30px;
-  &:before {
-    content: "";
-    display: inline-block;
-    width: 12px;
-    height: 12px;
-    border-radius: 50%;
-    margin-right: 5px;
-    background-color: #000;
-    margin-bottom: 5px;
-  }
   ${pMedia.sm`
-    margin-left: 0px;
     display: block;
   `}
 `
